--- React/react_todo_list/src/interfaces/reactModalMethods.ts
@@ -1,3 +1,0 @@
-export interface ReactModalMethods {
-  showDeleteModal: (id: number) => void;
-}
